Move wait functions to be reusable
