+++ integration/test/basic/basic_test.go
@@ -1,23 +1,18 @@
-// +build k8srequired
-
 package basic
 
 import (
 	"fmt"
-	"log"
 	"testing"
-	"time"
 
 	"github.com/spf13/afero"
 	"k8s.io/helm/pkg/helm"
 
-	"github.com/cenkalti/backoff"
 	"github.com/giantswarm/apprclient"
-	"github.com/giantswarm/e2e-harness/pkg/framework"
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 
+	"github.com/giantswarm/chart-operator/integration/release"
 	"github.com/giantswarm/chart-operator/integration/templates"
 )
 
@@ -43,7 +38,7 @@ func TestChartLifecycle(t *testing.T) {
 		t.Fatalf("could not install %q %v", cr, err)
 	}
 
-	err = waitForReleaseStatus(gsHelmClient, testRelease, "DEPLOYED")
+	err = release.WaitForReleaseStatus(gsHelmClient, testRelease, "DEPLOYED")
 	if err != nil {
 		t.Fatalf("could not get release status of %q %v", testRelease, err)
 	}
@@ -56,7 +51,7 @@ func TestChartLifecycle(t *testing.T) {
 		t.Fatalf("could not update %q %v", cr, err)
 	}
 
-	err = waitForReleaseVersion(gsHelmClient, testRelease, "5.6.0")
+	err = release.WaitForReleaseVersion(gsHelmClient, testRelease, "5.6.0")
 	if err != nil {
 		t.Fatalf("could not get release version of %q %v", testRelease, err)
 	}
@@ -69,7 +64,7 @@ func TestChartLifecycle(t *testing.T) {
 		t.Fatalf("could not delete %q %v", cr, err)
 	}
 
-	err = waitForReleaseStatus(gsHelmClient, testRelease, "DELETED")
+	err = release.WaitForReleaseStatus(gsHelmClient, testRelease, "DELETED")
 	if !helmclient.IsReleaseNotFound(err) {
 		t.Fatalf("%q not succesfully deleted %v", testRelease, err)
 	}
@@ -127,53 +122,5 @@ func updateChartOperatorResource(helmClient *helmclient.Client, releaseName stri
 		return microerror.Mask(err)
 	}
 
-	return nil
-}
-
-func waitForReleaseStatus(gsHelmClient *helmclient.Client, release string, status string) error {
-	operation := func() error {
-		rc, err := gsHelmClient.GetReleaseContent(release)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-		if rc.Status != status {
-			return microerror.Maskf(releaseStatusNotMatchingError, "waiting for %q, current %q", status, rc.Status)
-		}
-		return nil
-	}
-
-	notify := func(err error, t time.Duration) {
-		log.Printf("getting release status %s: %v", t, err)
-	}
-
-	b := framework.NewExponentialBackoff(framework.ShortMaxWait, framework.LongMaxInterval)
-	err := backoff.RetryNotify(operation, b, notify)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-	return nil
-}
-
-func waitForReleaseVersion(gsHelmClient *helmclient.Client, release string, version string) error {
-	operation := func() error {
-		rh, err := gsHelmClient.GetReleaseHistory(release)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-		if rh.Version != version {
-			return microerror.Maskf(releaseVersionNotMatchingError, "waiting for %q, current %q", version, rh.Version)
-		}
-		return nil
-	}
-
-	notify := func(err error, t time.Duration) {
-		log.Printf("getting release version %s: %v", t, err)
-	}
-
-	b := framework.NewExponentialBackoff(framework.ShortMaxWait, framework.LongMaxInterval)
-	err := backoff.RetryNotify(operation, b, notify)
-	if err != nil {
-		return microerror.Mask(err)
-	}
 	return nil
 }
